pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "palantir-stock"
version = "0.1.0"
description = "웹 검색 기반 기업 정보 수집 및 주식 데이터 분석 에이전트"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "Palantir Stock Team" }
]

dependencies = [
    # LLM & Agent Framework
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "openai>=1.0.0",

    # Search
    "google-search-results>=2.4.0",  # SerpAPI
    "tavily-python>=0.3.0",

    # Palantir Foundry
    "foundry-platform-sdk>=0.8.0",

    # Graph Database
    "neo4j>=5.0.0",
    "chromadb>=0.4.0",

    # Stock Data
    "yfinance>=0.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # Utilities
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "rich>=13.0.0",
    "typer>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
palantir = [
    "palantir-mcp",  # Palantir MCP integration
]

[project.scripts]
ps = "src.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
